requirements.txt
black==22.10.0
boto3==1.26.7
botocore==1.29.7
click==8.1.3
dnspython==2.2.1
jmespath==1.0.1
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
pathspec==0.10.1
platformdirs==2.5.3
py4j==0.10.9.5
pymongo==4.3.2
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
